+++ src/rendering/engine-Vk/engine-impl/resources/buffer.h
@@ -16,7 +16,7 @@ namespace engine {
                    const vk::PhysicalDevice& physicalDevice,
 				   const vk::Queue& queueHandle,
 				   const vk::CommandPool& cmdPool,
-                   std::uint32_t size,
+                   std::uint64_t size,
                    vk::BufferUsageFlags flags,
                    vk::MemoryPropertyFlags memoryFlags);
             ~buffer();
